+++ oidc_provider/models.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
-from hashlib import md5
+import base64
+import binascii
+from hashlib import md5, sha256
 import json
 
 from django.db import models
@@ -24,17 +26,18 @@ JWT_ALGS = [
     ('RS256', 'RS256'),
 ]
 
+
 class Client(models.Model):
 
-    name = models.CharField(max_length=100, default='')
-    client_type = models.CharField(max_length=30, choices=CLIENT_TYPE_CHOICES, default='confidential', help_text=_(u'<b>Confidential</b> clients are capable of maintaining the confidentiality of their credentials. <b>Public</b> clients are incapable.'))
-    client_id = models.CharField(max_length=255, unique=True)
-    client_secret = models.CharField(max_length=255, blank=True, default='')
-    response_type = models.CharField(max_length=30, choices=RESPONSE_TYPE_CHOICES)
+    name = models.CharField(max_length=100, default='', verbose_name=_(u'Name'))
+    client_type = models.CharField(max_length=30, choices=CLIENT_TYPE_CHOICES, default='confidential', verbose_name=_(u'Client Type'), help_text=_(u'<b>Confidential</b> clients are capable of maintaining the confidentiality of their credentials. <b>Public</b> clients are incapable.'))
+    client_id = models.CharField(max_length=255, unique=True, verbose_name=_(u'Client ID'))
+    client_secret = models.CharField(max_length=255, blank=True, default='', verbose_name=_(u'Client SECRET'))
+    response_type = models.CharField(max_length=30, choices=RESPONSE_TYPE_CHOICES, verbose_name=_(u'Response Type'))
     jwt_alg = models.CharField(max_length=10, choices=JWT_ALGS, default='RS256', verbose_name=_(u'JWT Algorithm'))
-    date_created = models.DateField(auto_now_add=True)
+    date_created = models.DateField(auto_now_add=True, verbose_name=_(u'Date Created'))
 
-    _redirect_uris = models.TextField(default='', verbose_name=_(u'Redirect URI'), help_text=_(u'Enter each URI on a new line.'))
+    _redirect_uris = models.TextField(default='', verbose_name=_(u'Redirect URIs'), help_text=_(u'Enter each URI on a new line.'))
 
     class Meta:
         verbose_name = _(u'Client')
@@ -45,12 +48,14 @@ class Client(models.Model):
 
     def __unicode__(self):
         return self.__str__()
-    
+
     def redirect_uris():
         def fget(self):
             return self._redirect_uris.splitlines()
+
         def fset(self, value):
             self._redirect_uris = '\n'.join(value)
+
         return locals()
     redirect_uris = property(**redirect_uris())
 
@@ -61,16 +66,18 @@ class Client(models.Model):
 
 class BaseCodeTokenModel(models.Model):
 
-    user = models.ForeignKey(settings.AUTH_USER_MODEL)
-    client = models.ForeignKey(Client)
-    expires_at = models.DateTimeField()
-    _scope = models.TextField(default='')
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, verbose_name=_(u'User'))
+    client = models.ForeignKey(Client, verbose_name=_(u'Client'))
+    expires_at = models.DateTimeField(verbose_name=_(u'Expiration Date'))
+    _scope = models.TextField(default='', verbose_name=_(u'Scopes'))
 
     def scope():
         def fget(self):
             return self._scope.split()
+
         def fset(self, value):
             self._scope = ' '.join(value)
+
         return locals()
     scope = property(**scope())
 
@@ -78,22 +85,22 @@ class BaseCodeTokenModel(models.Model):
         return timezone.now() >= self.expires_at
 
     def __str__(self):
-        return u'{0} - {1} ({2})'.format(self.client, self.user.email, self.expires_at)
+        return u'{0} - {1}'.format(self.client, self.user.email)
 
     def __unicode__(self):
         return self.__str__()
-    
+
     class Meta:
         abstract = True
 
 
 class Code(BaseCodeTokenModel):
 
-    code = models.CharField(max_length=255, unique=True)
-    nonce = models.CharField(max_length=255, blank=True, default='')
-    is_authentication = models.BooleanField(default=False)
-    code_challenge = models.CharField(max_length=255, null=True)
-    code_challenge_method = models.CharField(max_length=255, null=True)
+    code = models.CharField(max_length=255, unique=True, verbose_name=_(u'Code'))
+    nonce = models.CharField(max_length=255, blank=True, default='', verbose_name=_(u'Nonce'))
+    is_authentication = models.BooleanField(default=False, verbose_name=_(u'Is Authentication?'))
+    code_challenge = models.CharField(max_length=255, null=True, verbose_name=_(u'Code Challenge'))
+    code_challenge_method = models.CharField(max_length=255, null=True, verbose_name=_(u'Code Challenge Method'))
 
     class Meta:
         verbose_name = _(u'Authorization Code')
@@ -102,14 +109,18 @@ class Code(BaseCodeTokenModel):
 
 class Token(BaseCodeTokenModel):
 
-    access_token = models.CharField(max_length=255, unique=True)
-    refresh_token = models.CharField(max_length=255, unique=True, null=True)
-    _id_token = models.TextField()
+    access_token = models.CharField(max_length=255, unique=True, verbose_name=_(u'Access Token'))
+    refresh_token = models.CharField(max_length=255, unique=True, null=True, verbose_name=_(u'Refresh Token'))
+    _id_token = models.TextField(verbose_name=_(u'ID Token'))
+
     def id_token():
+
         def fget(self):
             return json.loads(self._id_token)
+
         def fset(self, value):
             self._id_token = json.dumps(value)
+
         return locals()
     id_token = property(**id_token())
 
@@ -117,8 +128,22 @@ class Token(BaseCodeTokenModel):
         verbose_name = _(u'Token')
         verbose_name_plural = _(u'Tokens')
 
+    @property
+    def at_hash(self):
+        # @@@ d-o-p only supports 256 bits (change this if that changes)
+        hashed_access_token = sha256(
+            self.access_token.encode('ascii')
+        ).hexdigest().encode('ascii')
+        return base64.urlsafe_b64encode(
+            binascii.unhexlify(
+                hashed_access_token[:len(hashed_access_token) // 2]
+            )
+        ).rstrip(b'=').decode('ascii')
+
 
 class UserConsent(BaseCodeTokenModel):
+
+    date_given = models.DateTimeField(verbose_name=_(u'Date Given'))
 
     class Meta:
         unique_together = ('user', 'client')
@@ -126,7 +151,7 @@ class UserConsent(BaseCodeTokenModel):
 
 class RSAKey(models.Model):
 
-    key = models.TextField(help_text=_(u'Paste your private RSA Key here.'))
+    key = models.TextField(verbose_name=_(u'Key'), help_text=_(u'Paste your private RSA Key here.'))
 
     class Meta:
         verbose_name = _(u'RSA Key')
@@ -140,4 +165,4 @@ class RSAKey(models.Model):
 
     @property
     def kid(self):
-        return  u'{0}'.format(md5(self.key.encode('utf-8')).hexdigest() if self.key else '')
+        return u'{0}'.format(md5(self.key.encode('utf-8')).hexdigest() if self.key else '')